Nouvelle classe SPlanet
Contient tous les éléments concernant la planète qui vont prochainement
aller dans le serveur. La classe fille Planet contiendra elle juste les
fonctions graphiques nécessaires au rendu et au fonctionnement du
client.

+++ include/Player.h
@@ -28,7 +28,7 @@ class Player
 
         sf::Vector2f getRelativePosition();
         sf::Vector2f getAbsolutePosition();
-        int getAngle();
+        float getAngle();
         sf::Color getColor();
         std::string getName();
     protected:
